@@ -13,30 +13,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "6.1.2",
+    "@angular/common": "6.1.2",
+    "@angular/compiler": "6.1.2",
+    "@angular/core": "6.1.2",
+    "@angular/forms": "6.1.2",
+    "@angular/http": "6.1.2",
+    "@angular/platform-browser": "6.1.2",
+    "@angular/platform-browser-dynamic": "6.1.2",
+    "@angular/router": "6.1.2",
     "@swimlane/ngx-datatable": "^11.3.2",
     "angular2-toaster": "^5.0.1",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "ng-block-ui": "^1.0.6",
-    "ngx-modal-dialog": "^2.0.1",
+    "ngx-modal-dialog": "^3.0.0",
     "ngx-tabs": "0.0.13",
     "ngx-tooltip": "0.0.9",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "npm-check-updates": "^2.14.2",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.3",
+    "@angular/compiler-cli": "6.1.2",
+    "@angular/language-service": "6.1.2",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -48,9 +51,9 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.9.2"
   }
 }
